renamed folders to lower case only
