+++ NixProxyModel.cpp
@@ -1,0 +1,7 @@
+#include "NixProxyModel.hpp"
+
+NixProxyModel::NixProxyModel()
+{
+
+}
+
