+++ NixProxyModel.hpp
@@ -1,0 +1,11 @@
+#ifndef NIXPROXYMODEL_H
+#define NIXPROXYMODEL_H
+
+
+class NixProxyModel
+{
+public:
+    NixProxyModel();
+};
+
+#endif // NIXPROXYMODEL_H
