+++ infowidget/DescriptionPanel.cpp
