+++ infowidget/DescriptionPanel.hpp
