+++ infowidget/DescriptionPanel.ui
