+++ infowidget/InfoWidget.cpp
