+++ infowidget/MetaDataPanel.cpp
