+++ infowidget/MetaDataPanel.ui
