+++ infowidget/TagPanel.cpp
