+++ infowidget/TagPanel.hpp
